Fix that CaptivePortalApp cannot handle the SSL error of POST request
Current design cannot distinguish the SSL error is coming from
the subresources or the POST request. It will make the log-in
page stop working when user tapped the submit/login button if
there is an SSL error from the POST request.
Fix this issue by getting the URL of POST request and
comparing if the SSL error URL is the same as the URL of POST
request.

Bug: 147195202
Test: 1. Add some logs locally and build & push the apk.
      2. Find a URL which will sent a POST request when tapping
         the LOGIN.
      3. Make that URL to a short link so that when visiting the
         short link, it will redirect to the original URL.
      4. Set the short link to captive_portal_http_url &
         captive_portal_https_url so that NetworkMonitor will
         think it's a captive portal.
      5. Open the CaptivePortalLogin APP.
      6. Tap the LOGIN and check the main frame URL which is
         coming from the shouldInterceptRequest() is the same as
         the action attribute of the form.
Change-Id: I7c176bdf6c076ed7226a68e876839d6034605cec